+++ src/balancer/balancer.ts
@@ -5,14 +5,17 @@ import * as path from "path";
 import * as cluster from "cluster";
 import { URL } from "url";
 import * as http from "http";
-import * as process from "process";
+import { User } from "@/types/User";
 
 export class Balancer {
   serverIndex: number;
+  users: User[];
+  workers: { worker: cluster.Worker; port: number }[];
 
   constructor(port: number) {
     const cpus = os.availableParallelism();
-    const workers: { worker: cluster.Worker; port: number }[] = [];
+    this.workers = [];
+    this.users = [];
 
     this.serverIndex = 0;
 
@@ -20,11 +23,25 @@ export class Balancer {
       for (let i = 0; i < cpus; i++) {
         const PORT = port + (i + 1);
         const fork = cluster.default.fork({ PORT });
-        workers.push({ port: PORT, worker: fork });
+        this.workers.push({ port: PORT, worker: fork });
+        fork.on("message", (message) => {
+          if (!("users" in message) || !("pid" in message)) {
+            return;
+          }
+
+          this.workers.map(({ worker }) => {
+            if (worker.id === message.pid) {
+              return;
+            }
+            worker.send({ users: message.users });
+          });
+        });
       }
 
       const server = http.createServer((req, res) => {
-        const serverUrl = `http://localhost:${workers[this.serverIndex].port}`;
+        const serverUrl = `http://localhost:${
+          this.workers[this.serverIndex].port
+        }`;
         const parsedUrl = new URL(serverUrl);
 
         if (
@@ -35,8 +52,6 @@ export class Balancer {
         ) {
           return;
         }
-
-        console.log("balancer url", req.url);
 
         const proxyReq = http.request(
           {
@@ -57,7 +72,7 @@ export class Balancer {
 
         req.pipe(proxyReq);
 
-        this.serverIndex = (this.serverIndex + 1) % workers.length;
+        this.serverIndex = (this.serverIndex + 1) % this.workers.length;
       });
 
       server.listen(8080, () => {
@@ -68,6 +83,21 @@ export class Balancer {
         return;
       }
       const crud = new Crud(parseInt(process.env.PORT));
+
+      process.on("message", (message) => {
+        const msg = message as { users: User[] };
+        if (!msg.users) {
+          return;
+        }
+        crud.users = msg.users;
+      });
+
+      crud.onUserUpdate = (users) => {
+        if (!process.send) {
+          return;
+        }
+        process.send({ users, pid: process.pid });
+      };
     }
   }
 }
